Changed some paths and code to get the app working in other OSes
Changed around on the location of the iconPath in main.js to follow the documentation of electron-packager. This way it will show the icon for both built packages and when running the app from source regardles of OS. I've also added some build options for linux as well as removed the extensions for the icon settings since electron-packager manages this itself depending in which OS it is packaging for.