Made some more changes to the cross platform packaging
Forgot to differentiate between ia32 and x64 in robotjs. Made some additions to also rebuild the robotjs package with the correct architecture before packaging the different packages. rebuild and build still packages 64bit windows package.
build-all triggers 32bit rebuild of robotjs, then it makes packages for all 32bit platforms before it rebuilds robotjs for 64bit and then makes packages for all the 64bit platforms. Not sure if --platform=all is necessary, but atleast the option is there if anyone wants to try to build for those platforms. I've only gotten to test the win32, win64 and linux64 packages which all works.