Fix for possible shortcut issue.
Fix for possible issue when using the shortcut CTRL+ALT+F12, the string it compares against is "Connected - Click to Disconnect", not "Connected." so it will never be seen as connected and never do anything else than to reconnect on triggering the shortcut. This cuts of the remaining characters after the ninth character so that it's seen as "Connected" instead of the whole string from the button.