OBS Groups, sigh (#2674)
* fix: obs variable types

- $obsInputMonitorType: object -> text
- $obsIsConnected: text -> bool
- $obsIsRecording: text -> bool
- $obsIsStreaming: text -> bool
- $obsSceneItemEnabled: text -> bool
- $obsSceneItemName: number -> text

* moar metadata

* start of actual group caching

* chore: Implement full OBS group caching system

Either need to make recursive calls to OBS to determine whether certain
events were triggered inside of a group or inside of a scene.
Alternatively, need to cache info about groups and active scenes.

This approach takes the latter option. More memory usage, but less IPC
network usage, and better fault tolerances.

- Caches the OBS preview and program scenes.
- Caches the list of groups, along with the owning scenes.
- Provides for a consistent OBS_SCENE_ITEM_ENABLE_STATE_CHANGED event
- Still a bit more work to do in relation to this.

* Polish up OBS groups

* Include Manual triggering for the new variables

* cleanup a todo

Oddly, deleting the last scene containing a group yields a SceneRemoved
for the scene, followed by a SceneRemoved for the group. But both do
fire.

* I missed the whole point of this PR

because I did all of the actual work on the wrong branch. Anyways, it'll
actually include the $obsSceneItemName now.

---------

Co-authored-by: Erik Bigler <erikbigler@gmail.com>