Refactoring the auth facilities in IntegrationManager and AuthManager
Fixing a bug in the createToken method of oauth2-client that made it so token was always seen as non expired by having our own createToken method.
Adding the ability to pass options to the oauth2 client to be added with the requests. This allows for Integrations to work with APIs that require certain additional fields (For example Id and Secret in body as required by Tiltify)
Adding a refreshToken() method to IntegrationManager to allow integrations to manually request a refresh token id autoRefresh is disabled.
Adding a getAuth() method to IntegrationManager. Integrations should use this method to get the auth with a valid token to allow the IntegrationManager to autoRefresh the token.
Emitting a "token-refreshed" event when the token is refreshed by either method in case the integration needs to do something with it.
Renaming the oauth property to auth so it is consistent, since both appeared, and consistent with the type definition currently exposed to integrations.